+++ .dockerignore
@@ -1,1 +1,4 @@
 node_modules
+.cache
+.git
+*.md
