+++ Dockerfile.service
@@ -8,19 +8,20 @@ ARG APP_NAME
 ADD libs libs
 
 # Add app to container
-ADD apis/$APP_NAME apis/$APP_NAME
+ADD apps/$APP_NAME apps/$APP_NAME
 
 # Add root npm dependencies
 ADD yarn.lock .
 ADD package.json .
 
-WORKDIR /usr/src/apis/$APP_NAME
+# WORKDIR /usr/src/apps/$APP_NAME
 
-# Yarn install and resolve
-RUN yarn
+# # Yarn install and resolve
 
-WORKDIR /usr/src
+# WORKDIR /usr/src
 ADD config.development.json .
 
-WORKDIR /usr/src/apis/$APP_NAME
+WORKDIR /usr/src/apps/$APP_NAME
+RUN yarn install --no-progress
+
 ENTRYPOINT ["yarn", "run", "start:dev"]
