+++ apps/gateway/providers/google-drive/lib/util.spec.js
