+++ apps/gateway/public/javascripts/signout.js
