+++ apps/photos-api/package.json
@@ -23,27 +23,16 @@
   },
   "homepage": "https://github.com/Etimo/photo-garden#readme",
   "dependencies": {
-    "amqp10": "^3.5.3",
-    "amqplib": "^0.5.2",
     "config": "^1.0.0",
+    "db": "^1.0.0",
     "express": "^4.16.2",
     "express-promise-router": "^3.0.1",
-    "google-auth-library": "^0.12.0",
-    "googleapis": "^23.0.0",
-    "morgan": "^1.9.0",
-    "nodemon": "^1.17.2",
-    "pg": "^7.4.1",
     "request": "^2.83.0",
     "rhea": "^0.2.9",
     "stomp-client": "^0.9.0",
     "stompit": "^0.25.0"
   },
   "devDependencies": {
-    "chai": "^4.1.2",
-    "eslint": "^4.12.1",
-    "lint-staged": "^7.0.0",
-    "mocha": "^3.1.2",
-    "nodemon": "^1.14.9",
-    "prettier": "^1.9.2"
+    "nodemon": "^1.17.3"
   }
 }
