+++ docker-compose.yml
@@ -26,7 +26,7 @@ services:
       args:
         APP_NAME: gateway
     volumes:
-      - ${PWD}/apis:/usr/src/apis
+      - ${PWD}/apps:/usr/src/apps
       - ${PWD}/libs:/usr/src/libs
     environment:
       - PHOTO_GARDEN_PROVIDERS_GOOGLE_DRIVE_CLIENT_ID
@@ -40,10 +40,26 @@ services:
     ports:
       - "3000:3000"
 
+  photo-import:
+    build:
+      context: .
+      dockerfile: Dockerfile.service
+      args:
+        APP_NAME: photo-import
+    volumes:
+      - ${PWD}/apps:/usr/src/apps
+      - ${PWD}/libs:/usr/src/libs
+    depends_on:
+      - queue
+      - db
+    links:
+      - queue
+      - db
+
   # photos_api:
-  #   build: ./apis/photos
+  #   build: ./apps/photos
   #   volumes:
-  #     - "./apis/photos/:/usr/src/app"
+  #     - "./apps/photos/:/usr/src/app"
   #   depends_on:
   #     - db
   #     - queue
