--- lerna.json
@@ -1,5 +1,0 @@
-{
-  "lerna": "2.9.0",
-  "packages": ["libs/*", "apis/*"],
-  "version": "0.0.0"
-}
