+++ libs/db/package.json
@@ -1,14 +1,11 @@
 {
-  "name": "provider-user",
+  "name": "db",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "db": "^1.0.0"
+    "pg": "^7.4.1"
   }
 }
