+++ package.json
@@ -5,21 +5,22 @@
   },
   "homepage": "https://github.com/Etimo/photo-garden#readme",
   "workspaces": [
-    "apis/*",
+    "apps/*",
     "libs/*"
   ],
   "scripts": {
-    "prettier": "prettier --write \"**/*.{js,json,css,md}\" && git add"
+    "prettier": "prettier --write \"**/*.{js,json,css,md}\" && git add",
+    "update-docker-compose": "node update-docker-compose.js"
   },
   "pre-commit": [
     "prettier"
   ],
   "dependencies": {},
   "devDependencies": {
+    "handlebars": "^4.0.11",
     "husky": "^0.14.3",
     "lint-staged": "^7.0.0",
     "pre-commit": "^1.2.2",
-    "prettier": "^1.11.1",
-    "rollup": "^0.57.1"
+    "prettier": "^1.11.1"
   }
 }
